refactor: merge code into bct-patroller module, remove unused code
