--- modules/bct-patroller/client/newbie.html
@@ -1,133 +1,0 @@
-<h4>Patrol</h4>
-
-<div class="thread-data">
-  <div id="{{::post.id}}" ng-repeat="post in NewbieCtrl.posts track by post.id" class="post-block" ng-class="{ 'highlighted': post.highlighted, 'hidden': post.hidden, 'deleted': post._deleted, 'editing': post.id === NewbieCtrl.posting.post.id }">
-    <!-- Delete Post View -->
-    <div class="deleted" ng-if="post._deleted">Post Deleted</div>
-
-    <!-- Visible Post View -->
-    <div ng-if="!post._deleted">
-      <!-- Post Profile Section -->
-      <div class="post-user">
-        <a ui-sref="profile.posts({ username: post.user.username})">
-          <div class="user-avatar">
-            <img ng-style="NewbieCtrl.avatarHighlight(post.user.highlight_color)" ng-src="{{::post.avatar || $webConfigs.default_avatar}}">
-          </div>
-          <div class="user-role" ng-bind="post.user.role_name || 'user'"></div>
-        </a>
-      </div>
-
-      <!-- Post Body Section -->
-      <div class="post-content">
-        <!-- Post Title -->
-        <div class="post-title">
-          <span ng-if="post.user.online" class="online green"><i class="fa fa-circle"></i></span>
-          <span ng-if="!post.user.online" class="online green"><i class="fa fa-circle-o"></i></span>
-          <a class="username" data-balloon="{{post.user.role_name || 'User'}}" ui-sref="profile.posts({ username: post.user.username})" ng-style="NewbieCtrl.usernameHighlight(post.user.highlight_color)" ng-bind="::post.user.username"></a>
-          <div title="{{post.user.name}}" ng-if="post.user.name" class="display-name">
-            <span ng-bind-html="::post.user.name | truncate:33"></span>
-            <span class="hide-mobile">&nbsp;&mdash;&nbsp;</span>
-          </div>
-          <div class="timestamp">
-            <span ng-bind="::post.created_at | humanDate"></span>
-            <span ng-if="NewbieCtrl.showEditDate(post)" ng-bind-html="'&nbsp;&mdash;&nbsp;Edited ' + (post.updated_at | humanDate)"></span>
-          </div>
-
-          <ul class="post-action">
-            <li ng-if="NewbieCtrl.canDelete(post) && !post.deleted && post.position !== 1">
-              <a ng-href="#" data-balloon="Hide" ng-click="NewbieCtrl.openDeleteModal($index)">
-                <i class="fa fa-eye-slash"></i>
-              </a>
-            </li>
-            <li ng-if="NewbieCtrl.canDelete(post) && post.deleted">
-              <a ng-href="#" data-balloon="Unhide" ng-click="NewbieCtrl.openUndeleteModal($index)">
-                <i class="fa fa-eye"></i>
-              </a>
-            </li>
-            <li ng-if="NewbieCtrl.canPostLock(post) && !post.locked">
-              <a ng-href="#" data-balloon="Lock Post" ng-click="NewbieCtrl.lockPost(post)">
-                <i class="icon-epoch-lock"></i>
-              </a>
-            </li>
-            <li ng-if="NewbieCtrl.canPostLock(post) && post.locked">
-              <a ng-href="#" data-balloon="Unlock Post" ng-click="NewbieCtrl.unlockPost(post)">
-                <i class="icon-epoch-lock clicked"></i>
-              </a>
-            </li>
-            <li ng-if="NewbieCtrl.loggedIn() && post.user.id !== NewbieCtrl.user.id && (NewbieCtrl.reportControlAccess.reportPosts || NewbieCtrl.reportControlAccess.reportUsers)">
-              <a ng-href="#" data-balloon="Report" ng-click="NewbieCtrl.openReportModal(post)">
-                <i class="icon-epoch-flag"></i>
-              </a>
-            </li>
-            <li ng-if="NewbieCtrl.canUpdate(post)">
-              <a ng-href="#" data-balloon="Edit" ng-click="NewbieCtrl.loadEditor(post)">
-                <i class="icon-epoch-edit-pencil"></i>
-              </a>
-            </li>
-          </ul>
-          <div class="clear"></div>
-        </div>
-        <!-- Thread title -->
-        <div class="thread-title">
-          <span ng-repeat="breadcrumb in post.breadcrumbs" ng-if="breadcrumb.label" ng-class="{ active: $last }">
-            <a ng-if="!$last && breadcrumb.state" title="{{breadcrumb.label}}" ui-sref="{{breadcrumb.state}}({{breadcrumb.opts}})" ui-sref-opts="{ reload: true }" ng-bind-html="breadcrumb.label | truncate:30"></a>
-            <strong ng-if="!$last">&nbsp;/&nbsp;</strong>
-            <strong ng-if="$last"><a ui-sref="posts.data({ threadId: post.thread_id, start: post.position, '#': post.id })" ng-bind-html="::post.thread_title"></a></strong>
-          </span>
-        </div>
-        <!-- Post Body -->
-        <div class="post-body" post-processing="post.content.body_html || post.content.body" style-fix="true"></div>
-        <div ng-if="post.user.signature">
-          <div class="post-sig-border"></div>
-          <div class="post-signature" post-processing="post.user.signature" style-fix="true"></div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Hide Post Modal -->
-<modal show="NewbieCtrl.showDeleteModal">
-  <form name="$parent.form" class="css-form" novalidate>
-    <h3 class="thin-underline">Hide Post</h3>
-    <p>Are you sure you want to hide this post?</p>
-    <label ng-if="NewbieCtrl.canPostLockQuick(NewbieCtrl.deletePostIndex)">
-      <input type="checkbox" ng-model="NewbieCtrl.deleteAndLock" modal-focus="{{NewbieCtrl.showDeleteModal}}" />
-      Lock Post (poster cannot hide/edit this post)?
-    </label>
-    <!-- Hide Button -->
-    <div class="clear">
-      <button class="fill-row" ng-click="NewbieCtrl.deletePost()">
-        Hide Post
-      </button>
-    </div>
-  </form>
-</modal>
-
-<!-- Unhide Post Modal -->
-<modal show="NewbieCtrl.showUndeleteModal">
-  <form name="$parent.form" class="css-form" novalidate>
-    <h3 class="thin-underline">Unhide Post</h3>
-    <p>Are you sure you want to unhide this post?</p>
-    <!-- Unhide Button -->
-    <div class="clear">
-      <button class="fill-row" ng-click="NewbieCtrl.undeletePost()">
-        Unhide Post
-      </button>
-    </div>
-  </form>
-</modal>
-
-<!-- Purge Post Modal -->
-<modal show="NewbieCtrl.showPurgeModal">
-  <form name="$parent.form" class="css-form" novalidate>
-    <h3 class="thin-underline">Purge Post</h3>
-    <p>Are you sure you want to permanently delete this post?</p>
-    <!-- Save Button -->
-    <div class="clear">
-      <button class="fill-row" ng-click="NewbieCtrl.purgePost()">
-        Purge Post
-      </button>
-    </div>
-  </form>
-</modal>
