+++ modules/bct-patroller/client/patrol.html
