+++ modules/bct-patroller/db/index.js
@@ -1,5 +1,5 @@
 var path = require('path');
 
 module.exports = {
-  publicNewbiePosts: require(path.normalize(__dirname + '/publicNewbiePosts')),
+  patrol: require(path.normalize(__dirname + '/patrol')),
 };
