+++ modules/bct-patroller/index.js
@@ -1,13 +1,9 @@
 var path = require('path');
 var db = require(path.normalize(__dirname + '/db'));
 var routes = require(path.normalize(__dirname + '/routes'));
-// var permissions = require(path.normalize(__dirname + '/permissions'));
-// var authorization = require(path.normalize(__dirname + '/authorization'));
 
 module.exports =  {
   name: 'patroller',
-  // authorization: authorization,
-  // permissions: permissions,
   routes: routes,
   db: db
 };
