+++ modules/bct-patroller/routes/index.js
@@ -1,5 +1,5 @@
 var path = require('path');
 
 module.exports = [
-  require(path.normalize(__dirname + '/publicNewbiePosts')),
+  require(path.normalize(__dirname + '/patrol')),
 ];
