+++ modules/ept-posts/client/resource.js
@@ -52,10 +52,6 @@ var resource = ['$resource',
         url: '/api/posts/:id/purge',
         method: 'DELETE',
         params: { id: '@id' }
-      },
-      patrolPosts: {
-        method: 'GET',
-        url: '/api/posts/patrol'
       }
     });
   }
